Extract API versioning into dedicated core microcrate
